refactor: extract AppDestination and PlannerStore into dedicated files
